avoid calls of _mm256_* functions in the Vc implementation, use wrappers instead
The wrappers are all implemented in avx/intrinsics.h, simply forwarding
to the _mm256_ functions. This will allow to switch the AVX2 intrinsics
between actual AVX2 instructions and AVX instructions. This would be
more easy if compilers were nice, but the world is a cruel place...
(well, at least the commercial compiler world)