unittest: less explicit Vector<T> instantiation
the ULP code lead to the requirement of abs/frexp/ldexp and more to be
implemented before any unit test was compilable. That is not usable for
implementing a new target. This can be fixed by having the functions
only instantiated when used (via templates). There may still be a small
error here, but at least NEON stuff can be compiled now.

Signed-off-by: Matthias Kretz <kretz@kde.org>