simdize: only do "no transformation" for Category::None
Currently if the Category passed to ReplaceTypes was not specialized it
would simply do nothing. You'd then get a strange compile error
somewhere later in the user code. Not cool. Now it fails to compile on
instantiation of ReplaceTypes trying to derive from
The_simdization_for_the_requested_type_is_not_implemented. This at least
tries to give a hint what is going on, though it is hidden underneath a
lot of noise.

Signed-off-by: Matthias Kretz <kretz@kde.org>