Storage: rewrote the union/builtin/may_alias abstraction
The VectorMemoryUnion helper class was completely messed up at this
point. This commit introduces a complete rewrite and tries to find the
minimal changes needed to make it compile with GCC and clang.

The Storage abstraction now uses a third template argument to select the
implementation strategy instead of a huge pile of #ifdefs inside the
class definition.

Most importantly, the Storage abstraction is not built from an intrinsic
type parameter, but from the EntryType/ValueType and a number of entries
in the vector. The consequence is that Masks must use the T parameter to
declare their storage type. This is intended, as this is the only way to
properly work with vector builtins.

Signed-off-by: Matthias Kretz <kretz@kde.org>