AVX: let int_v::size() == 4
Up to now I had float_v::size() == int_v::size(). But without AVX2 the
width of integral SIMD instructions is only 128 bits. Therefore, the
natural size for AVX::int_v is 4 and not 8.

For now I disabled the AVX2 targets to make the transition to
int_v::size() == 4 a little easier.

gh-1

Signed-off-by: Matthias Kretz <kretz@kde.org>