Masks: use proxy object for mutable operator[]
The non-const subscript operator for Masks must return a type that is
assignable with a bool object. This is not always possible with an
lvalue reference. It is possible for Scalar, SSE, and AVX masks. Though
the bit representation for the different mask types is significantly
different. For MIC there is no way to use an lvalue reference unless the
Mask type would use a different bit representation than the native mask
register of the architecture. The same is true for AVX-512. Furthermore,
the combination of different MASK types in simd_mask_array leads to
non-uniform types in the underlying storage for a single subscript
operator in simd_mask_array.

As a consequence the return type of the Mask subscript operator must be
an rvalue proxy type. The Common::MaskEntry type implements the
necessary proxy but requires a setEntry callback in the Mask types.

Signed-off-by: Matthias Kretz <kretz@kde.org>